requirements.txt
# requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
kafka-python>=2.0.2
django-cors-headers>=4.3.0
channels>=4.0.0
websockets>=12.0
python-dotenv>=1.0.0
djangorestframework-simplejwt>=5.0.0